[REFACTOR][TIRX] Consolidate split host device stages
The host/device split flow already depends on annotation, extraction, and kernel-launch lowering running as one sequence, so exposing the three stages separately leaves an unnecessary public surface and makes their ordering easier to misuse.

This consolidates the stage implementations into the single SplitHostDevice pass, removes the old public wrappers and registrations, and updates the lowering pipelines and tests to exercise the one-pass API.
