+++ python/tvm/tirx/compilation_pipeline.py
@@ -48,9 +48,7 @@ def default_tir_pipeline():
                 tirx.transform.FP8ComputeLegalize(),
                 tirx.transform.VerifyMemory(),
                 tirx.transform.AnnotateEntryFunc(),
-                tirx.transform.AnnotateDeviceRegions(),
                 tirx.transform.SplitHostDevice(),
-                tirx.transform.LowerDeviceKernelLaunch(),
                 tirx.transform.MakePackedAPI(),
                 tirx.transform.FP8StorageLegalize(),
                 tirx.transform.BF16StorageLegalize(),
@@ -89,9 +87,7 @@ def tirx_pipeline():
                 tirx.transform.FP8ComputeLegalize(),
                 tirx.transform.VerifyMemory(),
                 tirx.transform.AnnotateEntryFunc(),
-                tirx.transform.AnnotateDeviceRegions(),
                 tirx.transform.SplitHostDevice(),
-                tirx.transform.LowerDeviceKernelLaunch(),
                 tirx.transform.MakePackedAPI(),
                 tirx.transform.FP8StorageLegalize(),
                 tirx.transform.BF16StorageLegalize(),
@@ -122,9 +118,7 @@ def trn_pipeline():
             tirx.transform.StmtSimplify(),
             tirx.transform.RemoveNoOp(),
             tirx.transform.AnnotateEntryFunc(),
-            tirx.transform.AnnotateDeviceRegions(),
             tirx.transform.SplitHostDevice(),
-            tirx.transform.LowerDeviceKernelLaunch(),
             tirx.transform.MakePackedAPI(),
         ]
         return tvm.ir.transform.Sequential(passes)(mod)
