--- tests/python/tirx-transform/test_tir_transform_annotate_device_regions.py
@@ -1,73 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-
-import tvm
-import tvm.testing
-from tvm.script import ir as I
-from tvm.script import tirx as T
-
-
-def test_annotate_thread_extent():
-    """Annotation inserted at the "thread_extent" attribute"""
-
-    @I.ir_module
-    class Before:
-        @T.prim_func(s_tir=True)
-        def main(A: T.Buffer(16, "float32")):
-            T.func_attr({"target": T.target("cuda", host="llvm")})
-            i = T.launch_thread("threadIdx.x", 16)
-            A[i] = 0.0
-
-    @I.ir_module
-    class Expected:
-        @T.prim_func(s_tir=True)
-        def main(A: T.Buffer(16, "float32")):
-            T.func_attr({"target": T.target("cuda", host="llvm")})
-            T.attr(T.target("cuda"), "target", 0)
-            i = T.launch_thread("threadIdx.x", 16)
-            A[i] = 0.0
-
-    After = tvm.tirx.transform.AnnotateDeviceRegions()(Before)
-    tvm.ir.assert_structural_equal(After, Expected)
-
-
-def test_annotate_device_scope():
-    """Annotation inserted at the "device_scope" attribute"""
-
-    @I.ir_module
-    class Before:
-        @T.prim_func(s_tir=True)
-        def main(A: T.Buffer(1, "float32")):
-            T.func_attr({"target": T.target("cuda", host="llvm")})
-            T.attr(0, "device_scope", 0)
-            A[0] = 0.0
-
-    @I.ir_module
-    class Expected:
-        @T.prim_func(s_tir=True)
-        def main(A: T.Buffer(1, "float32")):
-            T.func_attr({"target": T.target("cuda", host="llvm")})
-            T.attr(T.target("cuda"), "target", 0)
-            T.attr(0, "device_scope", 0)
-            A[0] = 0.0
-
-    After = tvm.tirx.transform.AnnotateDeviceRegions()(Before)
-    tvm.ir.assert_structural_equal(After, Expected)
-
-
-if __name__ == "__main__":
-    tvm.testing.main()
